Fix mypy, add tests for the masking change
